cocodjango/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

beautifulsoup4==4.12.3
chromadb==0.5.20
Django==5.0.3
django-countries==7.6.1
django-debug-toolbar==4.4.2
django-defender==0.9.8
django-environ==0.11.2
django-fast-ratelimit==9.0.1
django-phonenumber-field==8.0.0
django-simple-history==3.7.0
djangorestframework==3.14.0
drf-yasg==1.21.7
fitz==0.0.1.dev2
openai==1.42.0
openpyxl==3.1.5
pandas==2.2.2
phonenumbers==8.13.40
pillow==10.3.0
psutil==6.0.0
PyMySQL==1.1.0
pyotp==2.9.0
PyPDF2==3.0.1
pytesseract==0.3.10
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.2
sentence-transformers==3.3.1
tiktoken==0.7.0
